requirements.txt
torch>=2.6.0
torchvision>=0.17.0
torchaudio>=2.6.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
streamlit>=1.22.0 